@@ -1,6 +1,6 @@
 {
   "name": "components-helper",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "description": "Based on the docs to provide code prompt files for vue component library",
   "main": "lib/index.js",
   "module": "lib/index.es.js",
@@ -16,7 +16,8 @@
     "build": "rollup -c rollup.config.js",
     "test": "node test/index.js",
     "lint": "eslint . --fix --ext .ts,.js --ignore-pattern \"/lib/\"",
-    "prettier": "prettier --write \"src/*.ts\" --ignore-unknown"
+    "prettier": "prettier --write \"src/*.ts\" --ignore-unknown",
+    "prepare": "husky install"
   },
   "files": [
     "lib"
@@ -46,24 +47,22 @@
     "node": ">10.0.0"
   },
   "dependencies": {
-    "fast-glob": "^3.2.5"
+    "fast-glob": "^3.2.7"
   },
   "devDependencies": {
-    "@types/node": "^14.14.35",
-    "@typescript-eslint/eslint-plugin": "^4.19.0",
-    "@typescript-eslint/parser": "^4.19.0",
-    "eslint": "^7.22.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "husky": "^4.3.6",
-    "lint-staged": "^10.5.4",
-    "prettier": "^2.2.1",
-    "rollup": "^2.42.3",
+    "@types/node": "^16.11.12",
+    "@typescript-eslint/eslint-plugin": "^5.6.0",
+    "@typescript-eslint/parser": "^5.6.0",
+    "eslint": "^8.4.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "husky": "^7.0.4",
+    "lint-staged": "^12.1.2",
+    "prettier": "^2.5.1",
+    "rollup": "^2.61.0",
     "rollup-plugin-typescript2": "^0.30.0",
-    "typescript": "^4.2.3"
+    "typescript": "^4.5.2"
   },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
+  "lint-staged": {
+    "*.{ts,js}": ["eslint --fix", "prettier --write"]
   }
 }
