@@ -2,13 +2,13 @@ import ts from 'rollup-plugin-typescript2'
 
 const commonConf = {
   input: './src/index.ts',
-  external: ['fast-glob'],
-  plugins:[
+  external: ['fast-glob', 'fs', 'path'],
+  plugins: [
     ts({
       extensions: ['.ts'],
       tsconfig: './tsconfig.json',
     }),
-  ]
+  ],
 }
 
 const list = [
@@ -16,16 +16,16 @@ const list = [
     file: 'lib/index.es.js',
     format: 'es',
     exports: 'named',
+    compact: true,
   },
   {
     file: 'lib/index.js',
-    format: 'umd',
+    format: 'cjs',
     exports: 'named',
-    name: 'ComponentsHelper',
     compact: true,
   },
 ]
 
-const buildConf = options => Object.assign({}, commonConf, options)
+const buildConf = (options) => Object.assign({}, commonConf, options)
 
-export default list.map(output => buildConf({ output }))
+export default list.map((output) => buildConf({ output }))
